sni: fix CI failures — Windows recv/send portability and clang-format
Two classes of CI failure on the sni-gateway-modes branch:

1. Windows (msbuild + gcc-mingw): the steady-state and handshake socket I/O
   in sni_gateway_tls.c passed uint8_t* buffers and size_t lengths straight to
   recv()/send(). On Windows recv/send take (char *, int), so the uint8_t*
   arguments tripped -Werror=pointer-sign and the lengths mismatched the int
   parameter. Cast the buffers to char*/const char* and the lengths to int at
   all six call sites (four in sni_gateway_tls.c, two in sni_gateway_http.c).
   This is a no-op on POSIX (recv/send take void*/size_t there) and matches how
   OpenVPN already passes BLENZ() as the recv length elsewhere. The feature is
   still rejected at option-parse time on Windows; this only makes the code
   compile cleanly on that toolchain.

2. clang-format style check: re-ran clang-format v21.1.8 (the version pinned in
   .pre-commit-config.yaml) over the changed files. Only trailing-comment
   alignment and one argument-wrapping in options.c, socket.c/.h and
   sni_gateway_http.c needed adjusting.

No behavior change. Builds clean with gcc/clang --enable-werror; all 21 unit
test binaries pass, including under -fsanitize=address,undefined.

+++ src/openvpn/sni_gateway_http.c
@@ -207,7 +207,7 @@ sni_gw_http_check_and_consume_request(struct stream_buf *sb, const char *require
     const char *line_end = memchr(data, '\r', (size_t)request_len);
     if (!line_end)
     {
-        goto reject; /* cannot happen (CRLFCRLF found) but be defensive */
+        goto reject;                   /* cannot happen (CRLFCRLF found) but be defensive */
     }
     const char *path_start = data + 4; /* just past "GET " */
     const char *sp = memchr(path_start, ' ', (size_t)(line_end - path_start));
@@ -287,7 +287,7 @@ sni_gw_http_send_101(socket_descriptor_t sd)
 
     while (sent < total)
     {
-        ssize_t s = send(sd, p + sent, total - sent, MSG_NOSIGNAL);
+        ssize_t s = send(sd, p + sent, (int)(total - sent), MSG_NOSIGNAL);
         if (s > 0)
         {
             sent += (size_t)s;
@@ -329,7 +329,7 @@ sni_gw_http_server_accept_upgrade(socket_descriptor_t sd, const char *require_pa
     /* Blocking read: accumulate until CRLFCRLF or cap. */
     while (total < (int)sizeof(buf))
     {
-        ssize_t n = recv(sd, buf + total, (size_t)(sizeof(buf) - total), 0);
+        ssize_t n = recv(sd, buf + total, (int)(sizeof(buf) - total), 0);
         if (n < 0)
         {
             if (openvpn_errno() == EINTR)
